added the includes and lib fixed

+++ libft/ft_strlen.c
@@ -1,4 +1,6 @@
-int     ft_strlen(const char *s)
+#include "libft.h"
+
+size_t     ft_strlen(const char *s)
 {
         size_t  i;
